+++ packages/amplify_auth_cognito/example/lib/Widgets/UpdateUserAttribute.dart
@@ -13,6 +13,7 @@
 // permissions and limitations under the License.
 //
 
+import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_core/amplify_core.dart';
 import 'package:amplify_flutter/amplify.dart';
 import 'package:flutter/material.dart';
@@ -52,7 +53,7 @@ class _UpdateUserAttributeWidgetState extends State<UpdateUserAttributeWidget> {
   void _updateAttribute() async {
     try {
       var res = await Amplify.Auth.updateUserAttribute(
-        userAttributeKey: _keyController.text,
+        userAttributeKey: CognitoUserAttributeKey.parse(_keyController.text),
         value: _valueController.text,
       );
       if (res.nextStep.updateAttributeStep == 'CONFIRM_ATTRIBUTE_WITH_CODE') {
@@ -107,7 +108,8 @@ class _UpdateUserAttributeWidgetState extends State<UpdateUserAttributeWidget> {
                 Navigator.of(context).pushReplacement(
                   MaterialPageRoute(
                     builder: (context) => ConfirmUserAttribute(
-                      userAttributeKey: _keyController.text,
+                      userAttributeKey:
+                          CognitoUserAttributeKey.parse(_keyController.text),
                     ),
                   ),
                 );