+++ packages/amplify_auth_cognito/test/amplify_auth_cognito_confirmSignIn_test.dart
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-import 'package:amplify_auth_cognito/src/CognitoSignUp/cognito_user_attributes.dart';
+import 'package:amplify_auth_cognito/src/CognitoSignUp/cognito_user_attribute_key.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
@@ -40,7 +40,7 @@ void main() {
 
   AmplifyAuthCognito auth = AmplifyAuthCognito();
   TestWidgetsFlutterBinding.ensureInitialized();
-  const testAttributeKey = CognitoUserAttributes.email;
+  const testAttributeKey = CognitoUserAttributeKey.email;
   const testEmailValue = 'test@test.test';
   const testMetadataKey = 'key';
   const testMetaDataAttribute = 'val';
@@ -61,7 +61,7 @@ void main() {
             "signInStep": "DONE",
             "codeDeliveryDetails": {
               "deliveryMedium": "EMAIL",
-              "attributeName": testAttributeKey,
+              "attributeName": testAttributeKey.key,
               "destination": testEmailValue
             }
           }
@@ -76,7 +76,7 @@ void main() {
         nextStep: AuthNextSignInStep(
           codeDeliveryDetails: AuthCodeDeliveryDetails(
               deliveryMedium: "EMAIL",
-              attributeName: testAttributeKey,
+              attributeName: testAttributeKey.key,
               destination: testEmailValue),
           signInStep: "DONE",
         ));
@@ -96,8 +96,7 @@ void main() {
     expect(req['options']['clientMetadata'][testMetadataKey],
         equals(testMetaDataAttribute));
     expect(req['options']['userAttributes'], isInstanceOf<Map>());
-    expect(req['options']['userAttributes'][testAttributeKey],
-        equals(testEmailValue));
+    expect(req['options']['userAttributes']['email'], equals(testEmailValue));
   });
 
   test('confirmSignIn thrown PlatFormException results in AuthException',