+++ packages/amplify_auth_plugin_interface/lib/src/Attribute/AuthUserAttribute.dart
@@ -13,26 +13,28 @@
  * permissions and limitations under the License.
  */
 
+import 'package:amplify_auth_plugin_interface/amplify_auth_plugin_interface.dart';
+
 class AuthUserAttribute {
-  String userAttributeKey;
-  var value;
-  AuthUserAttribute.init(
-      {required this.userAttributeKey, required this.value}) {
-    this.userAttributeKey = userAttributeKey;
-    if (userAttributeKey != 'phone_number') {
-      try {
-        this.value = int.parse(value);
-      } on FormatException {}
-    }
-  }
+  final UserAttributeKey userAttributeKey;
+  late final Object value;
 
   /// Creates an object that holds the key and value for a user attribute.
-  AuthUserAttribute({required this.userAttributeKey, required this.value}) {}
+  AuthUserAttribute({
+    required this.userAttributeKey,
+    required String value,
+  }) {
+    Object? _value;
+    if (userAttributeKey.key != 'phone_number') {
+      _value = int.tryParse(value);
+    }
+    this.value = _value ?? value;
+  }
 
   // ignore: public_member_api_docs
   Map<String, dynamic> serializeAsMap() {
     final Map<String, dynamic> pendingRequest = <String, dynamic>{};
-    pendingRequest['userAttributeKey'] = userAttributeKey;
+    pendingRequest['userAttributeKey'] = userAttributeKey.key;
     pendingRequest['value'] = value;
     return pendingRequest;
   }