+++ packages/amplify_datastore/example/lib/models/HasOneModel.dart
@@ -13,7 +13,11 @@
 * permissions and limitations under the License.
 */
 
-// ignore_for_file: public_member_api_docs
+// NOTE: This file is generated and may not follow lint rules defined in your app
+// Generated files can be excluded from analysis in analysis_options.yaml
+// For more info, see: https://dart.dev/guides/language/analysis-options#excluding-code-from-analysis
+
+// ignore_for_file: public_member_api_docs, file_names, unnecessary_new, prefer_if_null_operators, prefer_const_constructors, slash_for_doc_comments, annotate_overrides, non_constant_identifier_names, unnecessary_string_interpolations, prefer_adjacent_string_concatenation, unnecessary_const, dead_code
 
 import 'ModelProvider.dart';
 import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
@@ -27,6 +31,8 @@ class HasOneModel extends Model {
   final String? _name;
   final String? _childID;
   final ChildModel? _child;
+  final TemporalDateTime? _createdAt;
+  final TemporalDateTime? _updatedAt;
 
   @override
   getInstanceType() => classType;
@@ -66,11 +72,26 @@ class HasOneModel extends Model {
     return _child;
   }
 
+  TemporalDateTime? get createdAt {
+    return _createdAt;
+  }
+
+  TemporalDateTime? get updatedAt {
+    return _updatedAt;
+  }
+
   const HasOneModel._internal(
-      {required this.id, required name, required childID, child})
+      {required this.id,
+      required name,
+      required childID,
+      child,
+      createdAt,
+      updatedAt})
       : _name = name,
         _childID = childID,
-        _child = child;
+        _child = child,
+        _createdAt = createdAt,
+        _updatedAt = updatedAt;
 
   factory HasOneModel(
       {String? id,
@@ -108,7 +129,12 @@ class HasOneModel extends Model {
     buffer.write("HasOneModel {");
     buffer.write("id=" + "$id" + ", ");
     buffer.write("name=" + "$_name" + ", ");
-    buffer.write("childID=" + "$_childID");
+    buffer.write("childID=" + "$_childID" + ", ");
+    buffer.write("createdAt=" +
+        (_createdAt != null ? _createdAt!.format() : "null") +
+        ", ");
+    buffer.write(
+        "updatedAt=" + (_updatedAt != null ? _updatedAt!.format() : "null"));
     buffer.write("}");
 
     return buffer.toString();
@@ -116,7 +142,7 @@ class HasOneModel extends Model {
 
   HasOneModel copyWith(
       {String? id, String? name, String? childID, ChildModel? child}) {
-    return HasOneModel(
+    return HasOneModel._internal(
         id: id ?? this.id,
         name: name ?? this.name,
         childID: childID ?? this.childID,
@@ -130,10 +156,22 @@ class HasOneModel extends Model {
         _child = json['child']?['serializedData'] != null
             ? ChildModel.fromJson(
                 new Map<String, dynamic>.from(json['child']['serializedData']))
+            : null,
+        _createdAt = json['createdAt'] != null
+            ? TemporalDateTime.fromString(json['createdAt'])
+            : null,
+        _updatedAt = json['updatedAt'] != null
+            ? TemporalDateTime.fromString(json['updatedAt'])
             : null;
 
-  Map<String, dynamic> toJson() =>
-      {'id': id, 'name': _name, 'childID': _childID, 'child': _child?.toJson()};
+  Map<String, dynamic> toJson() => {
+        'id': id,
+        'name': _name,
+        'childID': _childID,
+        'child': _child?.toJson(),
+        'createdAt': _createdAt?.format(),
+        'updatedAt': _updatedAt?.format()
+      };
 
   static final QueryField ID = QueryField(fieldName: "hasOneModel.id");
   static final QueryField NAME = QueryField(fieldName: "name");
@@ -164,6 +202,18 @@ class HasOneModel extends Model {
         isRequired: false,
         ofModelName: (ChildModel).toString(),
         associatedKey: ChildModel.ID));
+
+    modelSchemaDefinition.addField(ModelFieldDefinition.nonQueryField(
+        fieldName: 'createdAt',
+        isRequired: false,
+        isReadOnly: true,
+        ofType: ModelFieldType(ModelFieldTypeEnum.dateTime)));
+
+    modelSchemaDefinition.addField(ModelFieldDefinition.nonQueryField(
+        fieldName: 'updatedAt',
+        isRequired: false,
+        isReadOnly: true,
+        ofType: ModelFieldType(ModelFieldTypeEnum.dateTime)));
   });
 }