+++ packages/amplify_datastore/ios/Classes/types/model/FlutterAuthRule.swift
@@ -24,6 +24,7 @@ public struct FlutterAuthRule {
     private let groupClaim : String?
     private let groups : [String]?
     private let groupsField : String?
+    private let provider: String?
     private var operations : [String]?
     
     init(serializedData: [String: Any]) throws {
@@ -35,6 +36,7 @@ public struct FlutterAuthRule {
                 fieldName: "authStrategy",
                 desiredType: "String")
         }
+
         self.authStrategy = authStrategy
         
         self.ownerField = serializedData["ownerField"] as? String
@@ -46,6 +48,8 @@ public struct FlutterAuthRule {
         self.groups = serializedData["groups"] as? [String]
         
         self.groupsField = serializedData["groupsField"] as? String
+
+        self.provider = serializedData["provider"] as? String
 
         self.operations = serializedData["operations"] as? [String]
     }
@@ -62,6 +66,23 @@ public struct FlutterAuthRule {
                 return AuthStrategy.public
             default:
                 preconditionFailure("Could not create a AuthStrategy from \(authStrategyString)")
+        }
+    }
+
+    private func stringToAuthProvider(providerString: String?) -> AuthRuleProvider? {
+        switch providerString {
+        case "APIKEY":
+            return AuthRuleProvider.apiKey
+        case "OIDC":
+            return AuthRuleProvider.oidc
+        case "IAM":
+            return AuthRuleProvider.iam
+        case "USERPOOLS":
+            return AuthRuleProvider.userPools
+        case "FUNCTION":
+            return AuthRuleProvider.function
+        default:
+            return nil
         }
     }
     
@@ -81,7 +102,6 @@ public struct FlutterAuthRule {
     }
     
     public func convertToNativeAuthRule() -> AuthRule{
-        
         return AuthRule(
             allow: stringToAuthStrategy(authStrategyString: authStrategy),
             ownerField: ownerField,
@@ -89,6 +109,7 @@ public struct FlutterAuthRule {
             groupClaim: groupClaim,
             groups: groups ?? [String](),
             groupsField: groupsField,
+            provider: stringToAuthProvider(providerString: provider),
             operations: (operations)?.map {
                 stringToModelOperation(modelOperationString: $0)
             } ?? [ModelOperation]()