+++ packages/amplify_datastore/lib/amplify_datastore.dart
@@ -15,7 +15,8 @@
 
 import 'dart:async';
 
-import 'package:amplify_core/types/exception/AmplifyException.dart';
+import 'package:amplify_datastore/types/DataStoreHubEvents/DataStoreHubEvent.dart';
+import 'package:amplify_core/amplify_core.dart';
 import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
 import 'package:meta/meta.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
@@ -67,7 +68,7 @@ class AmplifyDataStore extends DataStorePluginInterface {
     _instance = instance as AmplifyDataStore;
   }
 
-  StreamController get streamController {
+  StreamController<DataStoreHubEvent> get streamController {
     return streamWrapper.datastoreStreamController;
   }