+++ packages/amplify_datastore/test/amplify_datastore_delete_test.dart
@@ -13,11 +13,12 @@
  * permissions and limitations under the License.
  */
 
+import 'package:amplify_core/amplify_core.dart';
 import 'package:amplify_datastore/amplify_datastore.dart';
 import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:amplify_core/test_utils/get_json_from_file.dart';
+import 'package:amplify_test/amplify_test.dart';
 import './test_models/Post.dart';
 import 'test_models/ModelProvider.dart';