+++ packages/amplify_storage_plugin_interface/lib/amplify_storage_plugin_interface.dart
@@ -16,7 +16,7 @@
 library amplify_storage_plugin_interface;
 
 import 'dart:async';
-import 'package:amplify_core/types/index.dart';
+import 'package:amplify_core/amplify_core.dart';
 
 import 'src/types.dart';
 export 'src/types.dart';